walk interface and reset previous state for member-ordering rules
This change introduces resetting the previous `IModifiers` state when
visiting an interface declaration.  Without doing so, subsequent
interface declarations would be conflated as if they were the same
interface when using the `variables-before-functions` rule option.

Version `1.2.0` did not have this problem and `2.1.0` introduced this
regression.  The `memberordering-method.test.ts` has been updated
appropriately to test for this regression and an additional test has
been provided to ensure that interface declarations should trigger
similarly to class declarations.