Add support for a `null` formatter
When using tslint as a library at bitHound, I would like to use the raw object as
the output.

Expose the raw output as well as add a null formatter that just returns
null (since in my case I will not be using the output).

I can live without the nullFormatter, but having access to the raw
output would be awesome for library users.