Stylish Formatter: output all file names
Currently the stylish formatter *only* prints the first fileName, and
then a list of lint errors from all files without specifying to which
file they belong.
After this fix, each time a new file is reached, a new line will be
added, and then the name of the file, so each group of lint failures
will be group with their fileName.