Added an abstract OptionallyTypedRule
Like `AbstractRule`, it provides abstract `apply*` members that must be
implemented.

*(errata: switched the order of checks in `linter.ts` because it felt
quicker having a property check before the function call)*

Also moved `isTypedRule` logic to a separate function, as it's no longer
unique to the `TypedRule` class. An alternative I'd considered but
rejected would be to make a member `isTyped(): this is IRule` on
`IRule`, but that would require classes not inheriting from
`AbstractRule` to also define their own `isTyped()`.

Resolves #2091.