Added "check-type-list" as option to whitespace rule
It'll now check whitespace around the operators in syntax lists (`|` and `&`).
Fixes #2818.

Also changed references to the const options to use the actual option variable and alphabetized the listings. Figured it'd be a little cleaner this way.
The `maxLength` for options was listed at 7 despite there originally being 8 options (any particular reason?), so I bumped it to 9 for this 9th option.