noUnusedVariableRule should apply the ignorePattern to imports
I'm not able to turn this rule on for my project because we use tsx with a custom namespace.  The namespace must be imported, but the sugar'ed code doesn't use it.  I tried applying the ignore-pattern to this namespace and it didn't work without this change.

adding a test case, doc fix, and easier fix that ajafff suggested