fix(noUnusedVariableRule): remove warning
Remove warning printed to console when a user of `tslint` is using the
`no-unused-variable` rule while also using the TypeScript compiler
with the `noUnusedLocals` and `noUnusedParameters` options set.

While it's not technically necessary to use the `noUnusedVariableRule`
rule with the TypeScript compiler options set, using this rule
is still useful when using the `--fix` option in `tslint` to
automatically remove unused variables.

Fixes #3219