New rule dead-type-branch
That looks at if-blocks and tries to flag dead branches. More specifically,
this searches conditions of if for some type-based tests (typeof, instanceof,
type guards in functions), and remembers which variables are looked at. If
one of these appears in the following branches and its type is understood
by TS as `never`, then flag the branch, as it's likely dead.