Fix misleading error message for no-implicit-dependencies
The error message would make false claims about dependencies not
appearing in package.json.  A grep on the intended package.json would
show the dependency, making one start to suspect some kind of nasty
interaction with other package.json files in the monorepo.  Fix the error
message to hint that the dependency in the package.json was just in the
wrong section.