Updated max-line-length ignore-pattern example (#3834)
The example uses:

```
class [a-zA-Z] implements 
```

but this only matches a class that has a single character as name. Even if it is an example, the rule should accept more than one character for the class name. I struggled with this ignore pattern in a project, because I didn't read the pattern itself completely. Therefore I want to give this feedback, such that other people don't have to search for the reason of a failing rule, like me.