Update orderedImports for standalone groups and source-order any
This change updates the refactored "group" based orderedImports
code to handle the cases fixed in the following previous PRs:

- #4374: fixed grouping for import-sources-order
- #3733: check all imports of same type are grouped

As part of this we removed the concept of ImportType as this
is now covered by the defined groups for sorting.

Moved the check for ensuring groups are not standing alone
into the code checkBlocksGroup() code where other group checks
are done.

Ex:
```
import './baa';
import './baz';

import './caa';  // this should fail
```

Also had to update several other test cases to include the failure
notice about import groups needing to be together.