Remove check for grouping imports together by group.
This check is already handled by the check for groups needing to be sorted.

The idea is that once the first block of imports is found that has
an import that is out of order, then all future blocks also have
blocks out of order.  So instead of flagging all blocks we just flag
the first block that is causing the issue and provide a fix that will
fix all blocks.

This makes the case for manually fixing the import ordering much more
clear because people can solve one item at a time.  The automattic
fixing is also more clear because we don't end up with multiple
failures that have the same fix.