[quotemark] Exclude some cases from backtick rule (#4693)
* [quotemark] Exclude some cases from backtick rule

This commit makes quotemark backtick ignore use strict declarations, enum members, lookup types, and strings containing octal escape sequences.

* [quotemark] Fix comment on use strict check call

I had copy pasted and forgotten to change it. This changes that comment.

* Revert unrelated change, fix octal escape sequence check

This commit makes it so that if a string has a literal backslash instead of an actual octal escape sequence, it is not flagged.